K_means Clustering
The algorithm gets input(X) in matrix format and a random value is chosen for k (it’s chosen as 5 in this illustration). The idea is to return a dictionary with centers, which are calculated by the algorithm, as key values and the data points from the input matrix X as its values. Intuitively it can be thought as: given a bunch of data points, find the cluster of points that are closely associated with each other. So there have to be some metric to link the data points together. K-means uses Euclidean distance to link the data points. 

How does the algorithm work?
1. First get a center from X at randomly chosen k value like: 
c = [tuple(pt) for pt in random.sample(list(X), k)]
Here, c is a list of tuples with each tuple consisting of 4 data points. That’s because each row of X matrix has four columns of values. In the algorithm implementation, I chose k=5, so c has 5 tuples with each tuple consisting of 4 data points.

2. The next step is to find the e distance from every point in X to the every tuple in the center. Then find the minimum distance, out of all possible distance values per data point, and get its index using np.argmin
The corresponding value from the center ‘c’ is retrieved using the index value from np.argmin. i.e., c[np.argmin] and stored in a dictionary. The corresponding data points, per every iteration, are stored in the form of an array to the dictionary. And those data points are the values for the dictionary key.

3. A new mean is calculated using the dictionary values.

4. Stopping criteria is introduced by checking whether the distinct values of this new mean equal the distinct values of the existing center ‘c’. Break if it does. 

5. Else: The new mean will become the new center. So ‘c’ gets new value.

6. Continue till the maximum number of iterations.