Cross Validation-know how
Why do we use cross validation?
1. For evaluating a predictive model’s performance: To understand the predictive performance of a model before testing it on the precious test set. So how do we measure the predictive performance? It’s possible to get a part of the training dataset, commonly referred as holdout or validation set and test the performance of the model on it.

2. For hyper-parameter tuning: We ought to optimize the process of finding the best value for a model’s hyper-parameters (depending on the model) as they affect the model’s complexity. How? From a set of values for each hyper-parameter, we train the model on the training data using every value of the tuning parameter. We assess the performance of the fitted model on the validation set and we can tabulate this for the set of values we have chosen to consider for the tuning parameter and choose the specific model complexity according to the error on the validation set. And assess the performance of the selected model on our test set. 

Question: Which subset of observation should I use as my validation set?
Answer: All of the data subsets, because we then will have an advantage of averaging the performance across these validation sets and avoid any sensitivity you might have with one specific choice of validation set.

3. For feature selection: We could calculate the training and validation errors with sets of features and select the set with, ideally, low training and validation errors.

4. For model selection: Let’s say we are considering the metric: accuracy to measure the performance of a classification model. After fitting the choice of models, we could pick the one that offers high accuracy performance.

Why train-test split is not enough?
Training accuracy: The problem with training a model on an entire dataset and testing on the entire dataset will not be accurate as the metric: training accuracy seems to reward overly complex model that’s unlikely to generalize to feature data and that is called over fitting the training data. 

Testing accuracy: An alternative procedure we came up with is called the train-test-split (usually 80%-20% split) where we train the model on training data and then its performance on the test data. It does not reward overly complex models and therefore can avoid the over fitting problem. However, the testing accuracy suffers from high variance, meaning the testing accuracy varies a lot depending on which set of observations happen to be in the testing set.

Cross validation generates an accurate estimate of the out-of-sample accuracy compared to train and test set split. Also, it uses the data more efficiently repeating values in the training set and not repeating values in the test set.

How does cross-validation work?
The technique takes the training data (only the training data) and splits it into different folds of train and validation datasets. Then it trains the model using the train subset and tests using the validation subset (both gotten from the training data), which is a set of values that is only used once for each fold. What it means is that a set of values used as test data for fold 1 will not be used as test data for fold 2 and so on. 
So the idea is to generate validation set out of the training data and average the estimation errors thereby predicting what to expect on the unseen test data. We are not touching the real test data throughout the process.